CMake: Also take over ROOT cflags as initial flags.
Needed for e.g. std=c++11 - we have to compile with the
same std as the ROOT libraries for ABI compatibility,
otherwise functionality is not guaranteed.

Fixes compilation with standard ROOT 6 (c++11 ABI is required)
and also with ROOT 5 when compiled with c++11.
Should also fix for other ABIs like c++14 (untested).