jAnalyzer: Add travis testing.
Builds ROOT 5.34.26, compiles jAnalyzer with gcc and clang against that.
We have to compile a full ROOT since travis is at Ubuntu 12.04 and there is no PPA.
Still fits perfectly in the 50 minutes limit for a minimal ROOT build.

jAnalyzer itself not tested yet, since cmake does not allow to install (yet).